se reorganizan componentes en cada carpeta como recomendacion de parte en presentación.
