+++ src/App.jsx
@@ -8,7 +8,7 @@ import PerfilMascota from './pages/PerfilMascota';
 import ServiciosTurnos from './pages/ServiciosTurnos';
 //import Especialistas from './pages/Especialistas';
 import AdministracionPanel from './pages/AdministracionPanel';
-import ValidacionUsuario from './components/ValidaciónUsuario';
+import ValidacionUsuario from './components/ComponentsHelpers/ValidacionUsuario';
 import Error404 from './pages/Error404';
 import SobreNosotros from './pages/SobreNosotros';
