+++ src/components/General/ButtonOption.jsx
@@ -1,7 +1,7 @@
 // src/components/ButtonComponent.js
 import { Button} from 'react-bootstrap';
 import { useNavigate } from 'react-router-dom';
-import "../css/inicio.css"
+import "../../css/inicio.css"
 
 function ButtonOption({ route, text }) {
