+++ src/main.jsx
@@ -3,10 +3,10 @@ import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import "bootstrap/dist/css/bootstrap.min.css"
 import "bootstrap/dist/js/bootstrap.bundle.js"
-import Navigationbar from './components/Navigationbar.jsx'
-import FooterBar from './components/FooterBar.jsx'
 import { Container } from 'react-bootstrap'
 import { BrowserRouter } from 'react-router-dom'
+import Navigationbar from './components/General/Navigationbar.jsx'
+import FooterBar from './components/General/FooterBar.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
